Move location finder filters and map to separate block
