+++ modules/custom/location_finder/src/Plugin/Block/LocationFinder.php
@@ -3,9 +3,6 @@
 namespace Drupal\location_finder\Plugin\Block;
 
 use Drupal\Core\Block\BlockBase;
-use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
-use Drupal\openy_socrates\OpenySocratesFacade;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Provides a block with location finder.
@@ -16,45 +13,13 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  *   category = @Translation("Paragraph Blocks")
  * )
  */
-class LocationFinder extends BlockBase implements ContainerFactoryPluginInterface {
-
-  /**
-   * Openy Socrates Facade.
-   *
-   * @var OpenySocratesFacade
-   */
-  protected $socrates;
-
-  /**
-   * {@inheritdoc}
-   */
-  public function __construct(array $configuration, $plugin_id, $plugin_definition, OpenySocratesFacade $socrates) {
-    parent::__construct($configuration, $plugin_id, $plugin_definition);
-    $this->socrates = $socrates;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
-    return new static(
-      $configuration,
-      $plugin_id,
-      $plugin_definition,
-      $container->get('socrates')
-    );
-  }
+class LocationFinder extends BlockBase {
 
   /**
    * {@inheritdoc}
    */
   public function build() {
     return [
-      [
-        '#type' => 'openy_map',
-        '#show_controls' => TRUE,
-        '#element_variables' => $this->socrates->getLocationPins(),
-      ],
       [
         '#type' => 'view',
         '#name' => 'locations',
